delete copies and import instead
