+++ Cargo-component.lock
@@ -1,0 +1,3 @@
+# This file is automatically generated by cargo-component.
+# It is not intended for manual editing.
+version = 1
