+++ components/protocol-wasm/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+warg-protocol = { workspace = true }
 warg-crypto = { workspace = true }
 warg-transparency = { workspace = true }
 thiserror = { workspace = true }
