--- components/protocol-wasm/build.rs
@@ -1,30 +1,0 @@
-use std::{env, io::Result, path::PathBuf};
-
-fn main() -> Result<()> {
-    let warg_proto = PathBuf::from("src/warg.proto");
-    let proto_files = vec![warg_proto];
-    let root = PathBuf::from("src/");
-
-    // Tell cargo to recompile if any of these proto files are changed
-    for proto_file in &proto_files {
-        println!("cargo:rerun-if-changed={}", proto_file.display());
-    }
-
-    let descriptor_path = PathBuf::from(env::var("OUT_DIR").unwrap()).join("proto_descriptor.bin");
-
-    prost_build::Config::new()
-        // Save descriptors to file
-        .file_descriptor_set_path(&descriptor_path)
-        // Override prost-types with pbjson-types
-        .compile_well_known_types()
-        .extern_path(".google.protobuf", "::pbjson_types")
-        // Generate prost structs
-        .compile_protos(&proto_files, &[root])?;
-
-    let descriptor_set = std::fs::read(descriptor_path)?;
-    pbjson_build::Builder::new()
-        .register_descriptors(&descriptor_set)?
-        .build(&[".warg"])?;
-
-    Ok(())
-}
