+++ components/protocol-wasm/src/lib.rs
@@ -3,17 +3,18 @@ use std::str::FromStr; // 0.4.15
 
 use bindings::protocol;
 struct Component;
-pub mod operator;
-pub mod package;
-mod proto_envelope;
-pub mod registry;
-mod serde_envelope;
+// pub mod operator;
+// pub mod package;
+// mod proto_envelope;
+// pub mod registry;
+// mod serde_envelope;
 
 pub use semver::{Version, VersionReq};
 
 use anyhow::Error;
-pub use proto_envelope::{ProtoEnvelope, ProtoEnvelopeBody};
-pub use serde_envelope::SerdeEnvelope;
+// pub use proto_envelope::{ProtoEnvelope, ProtoEnvelopeBody};
+// pub use serde_envelope::SerdeEnvelope;
+use warg_protocol::{package, proto_envelope::{ProtoEnvelope, ProtoEnvelopeBody}, SerdeEnvelope};
 use warg_crypto::{signing, Decode};
 
 /// Represents information about a registry package.
@@ -44,27 +45,27 @@ impl PackageInfo {
 }
 
 /// Types for converting to and from protobuf
-pub mod protobuf {
-    #![allow(clippy::all)]
-    // Generated by [`prost-build`]
-    include!(concat!(env!("OUT_DIR"), "/warg.rs"));
-    // Generated by [`pbjson-build`]
-    include!(concat!(env!("OUT_DIR"), "/warg.serde.rs"));
+// pub mod protobuf {
+//     #![allow(clippy::all)]
+//     // Generated by [`prost-build`]
+//     include!(concat!(env!("OUT_DIR"), "/warg.rs"));
+//     // Generated by [`pbjson-build`]
+//     include!(concat!(env!("OUT_DIR"), "/warg.serde.rs"));
 
-    pub fn prost_to_pbjson_timestamp(timestamp: prost_types::Timestamp) -> pbjson_types::Timestamp {
-        pbjson_types::Timestamp {
-            seconds: timestamp.seconds,
-            nanos: timestamp.nanos,
-        }
-    }
+//     pub fn prost_to_pbjson_timestamp(timestamp: prost_types::Timestamp) -> pbjson_types::Timestamp {
+//         pbjson_types::Timestamp {
+//             seconds: timestamp.seconds,
+//             nanos: timestamp.nanos,
+//         }
+//     }
 
-    pub fn pbjson_to_prost_timestamp(timestamp: pbjson_types::Timestamp) -> prost_types::Timestamp {
-        prost_types::Timestamp {
-            seconds: timestamp.seconds,
-            nanos: timestamp.nanos,
-        }
-    }
-}
+//     pub fn pbjson_to_prost_timestamp(timestamp: pbjson_types::Timestamp) -> prost_types::Timestamp {
+//         prost_types::Timestamp {
+//             seconds: timestamp.seconds,
+//             nanos: timestamp.nanos,
+//         }
+//     }
+// }
 
 /// Helper module for serializing and deserializing timestamps.
 ///
@@ -137,6 +138,7 @@ fn perm_binding(perm: &package::model::Permission) -> protocol::Perm {
     match perm {
         &package::Permission::Release => protocol::Perm::Release,
         &package::Permission::Yank => protocol::Perm::Yank,
+        &_ => protocol::Perm::Release,
     }
 }
