+++ crates/protocol/src/registry.rs
@@ -124,7 +124,7 @@ impl AsRef<[u8]> for LogId {
 
 #[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(transparent)]
-pub struct RecordId(DynHash);
+pub struct RecordId(pub DynHash);
 
 impl RecordId {
     pub fn algorithm(&self) -> HashAlgorithm {
